requirements.txt
speedtest-cli
smbprotocol
python-dotenv
dropbox
